Long long variabler + udprint fix
Variabler til udprint af stats er lavet til long long, og har lavet en variabel "failed runs" til udskrift af hvor mange gange en algoritme fejlede

+++ src/algorithms/bfs.h
@@ -8,6 +8,7 @@
 #include "../utility/misc.h"
 
 bool bfs_is_valid(int height, int width, int grid[height][width], node current, node target);
-int bfs(int height, int width, int grid[height][width], long* direction_switches,node target_t,node start,long* total_tiles,node path[height * width],bool lastcase, long* elapsed_bfs);
+int bfs(int height, int width, int grid[height][width], long long* direction_switches,node target_t,node start,
+    long long* total_tiles,node path[height * width],bool lastcase, long long* elapsed_bfs);
 
 #endif //AUTOMATED_WAREHOUSE_BFS_H
